sleepit_unix_test: fix race when waiting for process to be ready
We were telling sleepit to sleep for 10ms, and checking each 10ms for
sleepit to print "sleepit: ready" before sending the signal.

So sometimes we were triggering a race condition, where sleepit consumed
its 10ms and terminated before the wait for the "sleepit: ready" was over.

To avoid this, it should be enough to check for "sleepit: ready" each 1ms
instead of each 10ms.